add aggregations helper
