+++ src/aggregations/bucket/multi_terms.ts
@@ -1,11 +1,11 @@
 import type {
-	AggregationFieldResult,
 	AppendSubAggs,
 	ElasticsearchIndexes,
 	SearchRequest,
 	TypeOfField,
 } from "../../";
 import type { PrettyArray, SeparatedString } from "../../types/helpers";
+import type { AggregationFieldResult } from "../helpers";
 
 /**
  * @see https://www.elastic.co/docs/reference/aggregations/search-aggregations-bucket-multi-terms-aggregation
